# flate: Improve speed in big stateless blocks.