--- a/.github/workflows/release.yml
+++ b/.github/workflows/release.yml
@@ -21,7 +21,7 @@ jobs:
           go-version: 1.19.x
       -
         name: install garble
-        run: go install mvdan.cc/garble@v0.7.1
+        run: go install mvdan.cc/garble@v0.7.2
       -
         name: Run GoReleaser
         uses: goreleaser/goreleaser-action@v2